🛂 Implement security, add extra features (#29)
* 28: Add JWT creation

* 28: Implement main logic for authentication

* 28: Implement endpoint to retrieve cars count

* 28: Update Postman collection

* 28: Try to fix CORS

* 28: Add Angular project to allowed origins in CORS

* 28: Implement endpoint to create manager; add role claim to JWT

* 28: Implement endpoint to retrieve random manager

* 28: Implement endpoint to retrieve page of users

* 28: Implement retrieving users and orders count

* 28: Update Postman collection

* 28: Implement status changing

* 28: Add endpoint to retrieve info about logged-in user

* 28: Update Postman collection

* 28: Fix mapping on car update

* 28: Implement order cancelling, fix car deletion

* add email sender to create order service

* 28: implement car deletion

---------

+++ car-dealership-system-api/src/main/java/com/gigateam/cardealershipsystemapi/service/impl/DefaultCarService.java
@@ -2,16 +2,21 @@ package com.gigateam.cardealershipsystemapi.service.impl;
 
 
 import com.gigateam.cardealershipsystemapi.common.dto.car.CarDto;
+import com.gigateam.cardealershipsystemapi.common.dto.order.OrderDto;
 import com.gigateam.cardealershipsystemapi.domain.Car;
 import com.gigateam.cardealershipsystemapi.domain.CarStatus;
+import com.gigateam.cardealershipsystemapi.exception.BadRequestException;
 import com.gigateam.cardealershipsystemapi.exception.NotFoundException;
 import com.gigateam.cardealershipsystemapi.repository.CarRepository;
 import com.gigateam.cardealershipsystemapi.service.CarService;
+import com.gigateam.cardealershipsystemapi.service.OrderService;
 import com.gigateam.cardealershipsystemapi.service.mapper.CarMapper;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Lazy;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -21,7 +26,13 @@ public class DefaultCarService implements CarService {
 
   private final CarRepository repository;
   private final CarMapper carMapper;
+  private OrderService orderService;
 
+  @Lazy
+  @Autowired
+  void setOrderService(OrderService orderService) {
+    this.orderService = orderService;
+  }
 
   @Override
   public Optional<CarDto> getCarById(Long id) {
@@ -47,6 +58,7 @@ public class DefaultCarService implements CarService {
 
     Car entity = carMapper.toEntity(car);
     entity.setId(id);
+    entity.setStatus(car.getStatus());
 
     return repository.save(entity).getId();
   }
@@ -54,6 +66,15 @@ public class DefaultCarService implements CarService {
   @Override
   @Transactional
   public boolean deleteCarById(Long id) {
+    boolean isCarOrdered = orderService.getOrdersByCarId(id).stream() //TODO: add status 'DELETED'
+        .anyMatch(OrderDto::isNotCancelled);
+
+    if (isCarOrdered) {
+      throw new BadRequestException(String.format("Car with id: %d is ordered", id));
+    }
+
+    orderService.deleteOrdersByCarId(id);
+
     return repository.deleteCarById(id) > 0;
   }
 
@@ -82,6 +103,11 @@ public class DefaultCarService implements CarService {
   @Override
   public boolean carNotExistsById(Long carId) {
     return !carExistsById(carId);
+  }
+
+  @Override
+  public Long getCarsCount(String query) {
+    return repository.count(query);
   }
 
 }
